bangkit-deploy-model-stress-detector/requirements.txt
Flask
tensorflow
numpy
Pillow